Fix bugs in delete and post methods

+++ src/main/java/com/strelnikov/bugtracker/service/ProjectServiceImpl.java
@@ -1,11 +1,13 @@
 package com.strelnikov.bugtracker.service;
 
 import com.strelnikov.bugtracker.configuration.PlainAuthentication;
+import com.strelnikov.bugtracker.entity.Issue;
 import com.strelnikov.bugtracker.entity.Project;
 import com.strelnikov.bugtracker.entity.ProjectRoleType;
 import com.strelnikov.bugtracker.entity.User;
 import com.strelnikov.bugtracker.exception.ProjectNotFoundException;
 import com.strelnikov.bugtracker.repository.ProjectRepository;
+import com.strelnikov.bugtracker.repository.ProjectRoleRepository;
 import com.strelnikov.bugtracker.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -13,24 +15,37 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class ProjectServiceImpl implements ProjectService {
 
     private ProjectRepository projectRepository;
+    private ProjectRoleRepository projectRoleRepository;
     private UserRepository userRepository;
+    private IssueService issueService;
 
     @Autowired
-    public ProjectServiceImpl(ProjectRepository projectRepository, UserRepository userRepository) {
+    public ProjectServiceImpl(ProjectRepository projectRepository, ProjectRoleRepository projectRoleRepository, UserRepository userRepository, IssueService issueService) {
         this.projectRepository = projectRepository;
+        this.projectRoleRepository = projectRoleRepository;
         this.userRepository = userRepository;
+        this.issueService = issueService;
     }
 
 
     private User getCurrentUser() {
         final var userId = ((PlainAuthentication) SecurityContextHolder.getContext().getAuthentication()).getPrincipal();
         return userRepository.findById(userId).orElseThrow();
+    }
+
+    @Override
+    @Transactional
+    public Project create(Project project) {
+        project.setId(0L);
+        Project newProject = projectRepository.save(project);
+        final User currentUser = this.getCurrentUser();
+        currentUser.addProjectRole(newProject, ProjectRoleType.ADMIN);
+        return newProject;
     }
 
     @Override
@@ -52,19 +67,21 @@ public class ProjectServiceImpl implements ProjectService {
 
     @Override
     public Project findById(Long projectId) {
-        Optional<Project> result = projectRepository.findById(projectId);
-        Project project = null;
-        if (result.isPresent()) {
-            project = result.get();
-        }
-        else {
-            throw new ProjectNotFoundException(projectId);
-        }
-        return project;
+        return projectRepository.findById(projectId)
+                .orElseThrow(() -> new ProjectNotFoundException(projectId));
     }
 
     @Override
+    @Transactional
     public void deleteById(Long projectId) {
+        if (!projectRepository.existsById(projectId)) {
+            throw new ProjectNotFoundException(projectId);
+        }
+        List<Issue> issues = issueService.findByProjectId(projectId);
+        for (Issue issue : issues) {
+            issueService.deleteById(issue.getId());
+        }
+        projectRoleRepository.deleteAllByProjectId(projectId);
         projectRepository.deleteById(projectId);
     }
